[PI-23] Add shared model module
