+++ AppPackage/Package.swift
@@ -13,7 +13,8 @@ let package = Package(
                 "MakePromise",
                 "PromiseManagement",
                 "LoginFeature",
-                "SwiftUIHelper"
+                "SwiftUIHelper",
+                "SharedModel"
             ]
         ),
         .library(
@@ -83,6 +84,7 @@ let package = Package(
             name: "CalendarFeature",
             dependencies: [
                 "DesignSystem",
+                "SharedModel",
                 .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
                 .product(name: "Introspect", package: "SwiftUI-Introspect")
             ]
@@ -141,7 +143,10 @@ let package = Package(
         ),
         .target(
             name: "SwiftUIHelper",
-            dependencies: []
-        )
+            dependencies: [
+                "SharedModel"
+            ]
+        ),
+        .target(name: "SharedModel")
     ]
 )
