+++ AppPackage/Sources/CalendarFeature/Model.swift
@@ -1,5 +1,6 @@
 import Foundation
 import IdentifiedCollections
+import SharedModel
 
 enum WeekDay: CaseIterable, CustomStringConvertible {
     var description: String {
@@ -43,32 +44,6 @@ public enum CalendarForm {
         self = self == .default
             ? .list
             : .default
-    }
-}
-
-public enum PromiseType: Equatable {
-    case meal
-    case meeting
-    case trip
-    case etc
-}
-
-public struct Promise: Identifiable, Equatable {
-    public let id: UUID
-    public let type: PromiseType
-    public let date: Date
-    public let name: String
-
-    public init(
-        id: UUID = .init(),
-        type: PromiseType,
-        date: Date,
-        name: String
-    ) {
-        self.id = id
-        self.type = type
-        self.date = date
-        self.name = name
     }
 }
